Adjust parallel testing to class-level from method-level
